requirements.txt
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
configparser==6.0.0
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
flask-marshmallow==0.15.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.1
psycopg2==2.9.7
psycopg2-binary==2.9.7
PyMySQL==1.1.0
requests==2.31.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.6
WTForms==3.0.1
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
configparser==6.0.0
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
flask-marshmallow==0.15.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.1
psycopg2==2.9.7
psycopg2-binary==2.9.7
PyMySQL==1.1.0
requests==2.31.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.6
WTForms==3.0.1
